+++ Lib/fontParts/base/anchor.py
@@ -392,5 +392,5 @@ class BaseAnchor(
 
         Subclasses may override this method.
         """
-        self.x = normalizers.normalizeRounding(self.x)
-        self.y = normalizers.normalizeRounding(self.y)
+        self.x = normalizers.normalizeVisualRounding(self.x)
+        self.y = normalizers.normalizeVisualRounding(self.y)