+++ Lib/fontParts/fontshell/glyph.py
@@ -1,4 +1,5 @@
 import defcon
+import booleanOperations
 from fontParts.base import BaseGlyph
 from fontParts.base.errors import FontPartsError
 from fontParts.fontshell.base import RBaseObject
@@ -132,6 +133,17 @@ class RGlyph(RBaseObject, BaseGlyph):
         glyph = self.naked()
         contour = glyph[index]
         glyph.removeContour(contour)
+
+    def _removeOverlap(self, **kwargs):
+        if len(self):
+            contours = list(self)
+            for contour in contours:
+                for point in contour.points:
+                    if point.type == "qcurve":
+                        raise TypeError("fontshell can't removeOverlap for quadratics")
+            self.clear(contours=True, components=False,
+                       anchors=False, guidelines=False, image=False)
+            booleanOperations.union(contours, self.getPointPen())
 
     def _correctDirection(self, trueType=False, **kwargs):
         self.naked().correctContourDirection(trueType=trueType)